(fix): Inject dependency on MongoClient
Rather than constructing this dependency from the connection string
provided in a JSON file, the class requiring the dependency should
receive it as a constructor parameter.

This make the dependency both more explicit, and more configurable.

The method initiating the connection opaquely has been marked as deprecated.

An example script is added demonstrating the use of this lib.