(feat): Synchronise names with qualifiers
If a name in the "Title+Date"-sheet has a qualifier attached, such as:

   ACME Corp. [institution]

...then in addition to writing it to the 'persons'-dictionary, it is
also synced into the 'institution'-dictionary.

The possible identifiers can be configured using the
`_qualifiers`-property on the `ValueList`-class.

Note that this behaviour is only triggered when explicit syncing of the
'institutions'-collection is requested, and will _always_ sync into that
collection.